requirements.txt
# System dependencies:
# - libmagic (brew install libmagic on macOS, apt-get install libmagic1 on Ubuntu)

imapclient
python-dotenv
pinecone
pinecone-client[gRPC]
pinecone-plugin-assistant
pandas
openpyxl
xlrd
pytest
pytest-mock
pytest-email
jinja2
psycopg2-binary
litellm
pgvector
sqlalchemy>=2.0.0
pypdf
llama-index-core
markdown
fastapi
uvicorn
python-magic
requests

